requirements.txt
numpy<2
fastapi
uvicorn[standard]
pydantic
openai
PyYAML
jsonschema
python-multipart
langdetect
emoji
farm-haystack
pytest
pytest-asyncio
# Linters
pylint
psycopg2-binary
sqlalchemy
